Separate Authentication from Authorization. Use single Access Manager.
