--- access-manager-api/src/main/java/cwms/cda/spi/AccessManagerProvider.java
@@ -1,6 +1,0 @@
-package cwms.cda.spi;
-
-public interface AccessManagerProvider {
-    public String getName();
-    public CdaAccessManager create();
-}
