+++ access-manager-api/src/main/java/cwms/cda/spi/IdentityProvider.java
@@ -1,0 +1,21 @@
+package cwms.cda.spi;
+
+import java.security.Principal;
+
+import io.javalin.http.Context;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+
+public interface IdentityProvider {
+    /**
+     * Key used in OpenAPI definition to distinguish Auth types
+     * @return
+     */
+    String getName();
+    boolean canAuth(Context ctx);
+    Principal authenticate(Context ctx);
+    /**
+     * Define the OpenAPI V3 Security Scheme for this manager
+     * @return
+     */
+    SecurityScheme getScheme();
+}
