+++ cwms-data-api/src/main/java/cwms/cda/CdaAccessManager.java
@@ -1,0 +1,33 @@
+package cwms.cda;
+
+import java.util.Set;
+
+import cwms.cda.data.dao.AuthDao;
+import cwms.cda.data.dao.JooqDao;
+import cwms.cda.security.DataApiPrincipal;
+import io.javalin.core.security.AccessManager;
+import io.javalin.core.security.RouteRole;
+import io.javalin.http.Context;
+import io.javalin.http.Handler;
+
+public class CdaAccessManager implements AccessManager {
+
+    @Override
+    public void manage(Handler handler, Context ctx, Set<RouteRole> routeRoles) throws Exception {
+        DataApiPrincipal p = getApiPrincipal(ctx);
+        if (p == null) {
+            AuthDao authDao = AuthDao.getInstance(JooqDao.getDslContext(ctx),ctx.attribute(ApiServlet.OFFICE_ID));
+            authDao.prepareGuestContext(ctx);
+        } else {
+            AuthDao.isAuthorized(ctx,p,routeRoles);
+            AuthDao.prepareContextWithUser(ctx, p);
+        }
+        
+        handler.handle(ctx);
+    }
+
+    DataApiPrincipal getApiPrincipal(Context ctx) {
+        return (DataApiPrincipal)ctx.sessionAttribute("principal");
+    }
+
+}
