+++ cwms-data-api/src/test/java/cwms/cda/api/DownstreamLocationsGetControllerIT.java
@@ -17,7 +17,7 @@ import cwms.cda.data.dto.stream.Stream;
 import cwms.cda.data.dto.stream.StreamLocation;
 import cwms.cda.formatters.ContentType;
 import cwms.cda.formatters.Formats;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import fixtures.CwmsDataApiSetupCallback;
 import fixtures.TestAccounts;
 import static io.restassured.RestAssured.given;
