+++ cwms-data-api/src/test/java/cwms/cda/api/TurbineChangesControllerIT.java
@@ -26,7 +26,7 @@ package cwms.cda.api;
 
 import static cwms.cda.api.Controllers.OVERRIDE_PROTECTION;
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
