+++ cwms-data-api/src/test/java/cwms/cda/api/WaterContractControllerTestIT.java
@@ -62,7 +62,7 @@ import java.nio.charset.StandardCharsets;
 import java.time.ZoneId;
 import static cwms.cda.api.Controllers.*;
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
