+++ cwms-data-api/src/test/java/cwms/cda/api/WaterContractTypeControllerTestIT.java
@@ -27,7 +27,7 @@
 package cwms.cda.api;
 
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static java.lang.String.format;
 import static org.hamcrest.Matchers.equalTo;
