+++ cwms-data-api/src/test/java/cwms/cda/api/WaterPumpDisassociateControllerTestIT.java
@@ -61,7 +61,7 @@ import java.time.ZoneId;
 
 import static cwms.cda.api.Controllers.*;
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
