+++ cwms-data-api/src/test/java/cwms/cda/api/location/kind/GateChangeControllerTestIT.java
@@ -54,7 +54,7 @@ import usace.cwms.db.jooq.codegen.packages.CWMS_PROJECT_PACKAGE;
 
 import static cwms.cda.api.Controllers.*;
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
