+++ cwms-data-api/src/test/java/cwms/cda/api/location/kind/VirtualOutletControllerTestIT.java
@@ -45,7 +45,7 @@ import org.junit.jupiter.api.Test;
 import static cwms.cda.api.Controllers.FAIL_IF_EXISTS;
 import static cwms.cda.api.Controllers.METHOD;
 import static cwms.cda.data.dao.DaoTest.getDslContext;
-import static cwms.cda.security.KeyAccessManager.AUTH_HEADER;
+import static cwms.cda.security.KeyAccessManagerProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
